refactor: moved middlewares and routers folders
