+++ src/configs/app.ts
@@ -1,11 +1,11 @@
-import { AuthRouter } from '@src/modules/auth/routers/AuthRouter';
 import express from 'express';
 // express-async-errors must be imported to ensure that errors thrown
 // in async functions are handled by the error handler.
 // This packaged remove the need of using try/catch with next() in
 // async functions.
 import 'express-async-errors';
-import { UsersRouter } from '../modules/users/routers/UsersRouter';
+import { UsersRouter } from '../routers/UsersRouter';
+import { AuthRouter } from '../routers/AuthRouter';
 
 const app = express();
